# packages/api-spec/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Manhaj API",
    "version": "1.0.0",
    "description": "Offline-First Student Question Bank API"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "authToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-auth-token"
      }
    },
    "schemas": {
...
      "Attempt": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "questionId": {
            "type": "string",
            "format": "uuid"
          },
          "choiceId": {
            "type": "string",
            "format": "uuid"
          },
          "isCorrect": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
    "/student/profile": {
      "get": {
        "summary": "Get student profile",
        "tags": [
          "Student"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Student profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "termId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update student profile",
        "tags": [
          "Student"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "termId": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated"
          }
        }
      }
    },
    "/student/grades-with-terms": {
      "get": {
        "summary": "Get grades with their terms",
        "tags": [
          "Student"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Grades with terms",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "name": {
                        "type": "string"
                      },
                      "terms": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "format": "uuid"
                            },
                            "name": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/student/attempts/sync": {
      "post": {
        "summary": "Sync student attempts to server",
        "tags": [
          "Student"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "attempts"
                ],
                "properties": {
                  "attempts": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Attempt"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Attempts synced successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "synced": {
                            "type": "boolean"
                          },
                          "reason": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/student/questions/unsolved": {
      "get": {
        "summary": "Get unsolved questions for a lecture",
        "tags": [
          "Student"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "lectureId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unsolved questions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Question"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/student/flags": {
      "post": {
        "summary": "Flag a question",
        "tags": [
          "Student"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "questionId"
                ],
                "properties": {
                  "questionId": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Question flagged"
          }
        }
      }
    },
    "/student/flags/{questionId}": {
      "delete": {
        "summary": "Remove flag from question",
        "tags": [
          "Student"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "questionId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Flag removed"
          }
        }
      }
    },
    "/student/devices/register": {
      "post": {
        "summary": "Register an Expo push token",
        "tags": [
          "Student"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "pushToken",
                  "platform"
                ],
                "properties": {
                  "pushToken": {
                    "type": "string"
                  },
                  "platform": {
                    "type": "string",
                    "enum": [
                      "ios",
                      "android"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Device token registered"
          }
        }
      }
    },
    "/student/cases": {
      "get": {
        "summary": "Get all cases for the student",
        "tags": [
          "Student"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of cases"
          }
        }
      },
      "post": {
        "summary": "Create a case",
        "tags": [
          "Student"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "lectureId",
                  "category",
                  "title",
                  "content"
                ],
                "properties": {
                  "lectureId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "case",
                      "disease",
                      "drug"
                    ]
                  },
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "answer": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Case created"
          }
        }
      }
    },
    "/student/cases/{id}": {
      "patch": {
        "summary": "Update a case",
        "tags": [
          "Student"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "category": {
                    "type": "string",
                    "enum": [
                      "case",
                      "disease",
                      "drug"
                    ]
                  },
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "answer": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Case updated"
          }
        }
      },
      "delete": {
        "summary": "Delete a case",
        "tags": [
          "Student"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Case deleted"
          }
        }
      }
    },
    "/student/notes": {
      "get": {
        "summary": "Get all notes for the student",
        "tags": [
          "Student"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of notes"
          }
        }
      },
      "post": {
        "summary": "Create a note",
        "tags": [
          "Student"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "lectureId",
                  "type",
                  "content"
                ],
                "properties": {
                  "lectureId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "note",
                      "recurring_question"
                    ]
                  },
                  "content": {
                    "type": "string"
                  },
                  "sourceQuestionId": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Note created"
          }
        }
      }
    },
    "/student/notes/{id}": {
      "patch": {
        "summary": "Update a note",
        "tags": [
          "Student"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "note",
                      "recurring_question"
                    ]
                  },
                  "content": {
                    "type": "string"
                  },
                  "sourceQuestionId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Note updated"
          }
        }
      },
      "delete": {
        "summary": "Delete a note",
        "tags": [
          "Student"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Note deleted"
          }
        }
      }
    },
...
    "/student/tasks/{id}": {
      "patch": {
        "summary": "Update a task",
        "tags": [
          "Student"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "taskType": {
                    "type": "string",
                    "enum": [
                      "zekr",
                      "wird"
                    ]
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "pending",
                      "in_progress",
                      "done",
                      "missed"
                    ]
                  },
                  "zekrTask": {
                    "type": "object",
                    "nullable": true
                  },
                  "wirdTask": {
                    "type": "object",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated"
          }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "tags": [
          "Student"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task deleted"
          }
        }
      }
    },
...
    "/student/reviewables/due": {
      "get": {
        "summary": "Get due reviewable items for the student",
        "tags": [
          "Student"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of due reviewable items"
          }
        }
      }
    },
...
  },
...
}